Feat: userLectrue, Lecture 도메인 API 보안 설정 매니저 전용 API 보안 설정 개인용 API 타유저 접근 금지 보안 설정

+++ src/main/java/com/example/DoroServer/domain/userLecture/api/UserLectureApi.java
@@ -1,6 +1,7 @@
 package com.example.DoroServer.domain.userLecture.api;
 
 import com.example.DoroServer.domain.user.entity.User;
+import com.example.DoroServer.domain.user.entity.UserRole;
 import com.example.DoroServer.domain.userLecture.dto.FindAllTutorsRes;
 import com.example.DoroServer.domain.userLecture.dto.FindMyLecturesRes;
 import com.example.DoroServer.domain.userLecture.dto.SelectTutorReq;
@@ -9,9 +10,12 @@ import com.example.DoroServer.domain.userLecture.service.UserLectureService;
 import com.example.DoroServer.global.common.SuccessResponse;
 import java.util.List;
 
+import com.example.DoroServer.global.exception.BaseException;
+import com.example.DoroServer.global.exception.Code;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
@@ -24,17 +28,21 @@ import javax.validation.Valid;
 public class UserLectureApi {
 
     private final UserLectureService userLectureService;
-    @ApiOperation(value = "강사목록 조회",notes = "id에 lecture id 를 넣으면 해당 강의에 신청한 강사 목록을 불러옵니다.")
+    @Secured("ROLE_ADMIN")
+    @ApiOperation(value = "강사목록 조회",notes = "id에 lecture id 를 넣으면 해당 강의에 신청한 강사 목록을 불러옵니다. 매니저 전용 API")
     @GetMapping("/lectures/{id}")
-    public SuccessResponse findAllTutors(@PathVariable("id") Long id) {
+    public SuccessResponse findAllTutors(@PathVariable("id") Long id,@AuthenticationPrincipal User user) {
         List<FindAllTutorsRes> allTutorsResList = userLectureService.findAllTutors(id);
         return SuccessResponse.successResponse(allTutorsResList);
     }
     @ApiOperation(value = "내 강의 조회" ,notes = "id 에 user id 를 넣으면 해당 user 가 신청한 강의를 불러옵니다.")
     @GetMapping("/users/{id}")
-    public SuccessResponse findMyLectures(@PathVariable("id") Long id) {
-        List<FindMyLecturesRes> findMyLecturesResList = userLectureService.findMyLectures(id);
-        return SuccessResponse.successResponse(findMyLecturesResList);
+    public SuccessResponse findMyLectures(@PathVariable("id") Long id,@AuthenticationPrincipal User user) {
+        if(user.getId().equals(id) || user.getRole().equals(UserRole.ROLE_ADMIN)) {
+            List<FindMyLecturesRes> findMyLecturesResList = userLectureService.findMyLectures(id);
+            return SuccessResponse.successResponse(findMyLecturesResList);
+        }
+        throw new BaseException(Code.FORBIDDEN);
     }
     @ApiOperation(value = "강사 신청",notes = "id에 강의 id를 createTutorReq 에 강사 id와 직업을 넣으면 해당 강의에 강사 신청이 됩니다.")
     @PostMapping("/lectures/{id}")
@@ -42,11 +50,16 @@ public class UserLectureApi {
     //이미 신청한 user  일시 예외 출력
     public SuccessResponse createTutor(
             @PathVariable("id") Long id,
-            @RequestBody @Valid CreateTutorReq createTutorReq) {
-        Long userLectureId = userLectureService.createTutor(id, createTutorReq );
-
-        return SuccessResponse.successResponse(userLectureId + "is created");
+            @RequestBody @Valid CreateTutorReq createTutorReq,
+            @AuthenticationPrincipal User user
+    ) {
+        if (user.getId().equals(id)) {
+            Long userLectureId = userLectureService.createTutor(id, createTutorReq);
+            return SuccessResponse.successResponse(userLectureId + "is created");
+        }
+        throw new BaseException(Code.FORBIDDEN);
     }
+    @Secured("ROLE_ADMIN")
     @ApiOperation(value = "강사 선정",notes = "id에 강의 id를 selectTutorReq 에는 강사 id와 직업을 넣으면 해당 강사가 강의에 배정됩니다. ")
     @PatchMapping("/lectures/{id}")
     public SuccessResponse selectTutor(
@@ -57,9 +70,12 @@ public class UserLectureApi {
     }
     @ApiOperation(value="강사 신청 취소",notes = "id에 해당하는 신청내역을 삭제합니다.")
     @DeleteMapping("/lectures/{id}")
-    public SuccessResponse deleteUserLecture(@PathVariable("id") Long id){
-        Long deleteLectureID = userLectureService.deleteLecture(id);
-        return SuccessResponse.successResponse("deleted"+deleteLectureID);
+    public SuccessResponse deleteUserLecture(@PathVariable("id") Long id,@AuthenticationPrincipal User user){
+        if(user.getId().equals(id)){
+            Long deleteLectureID = userLectureService.deleteLecture(id);
+            return SuccessResponse.successResponse("deleted"+deleteLectureID);
+        }
+        throw new BaseException(Code.FORBIDDEN);
     }
 
 }
